Add hook for automatic compilation of bytecode functions
Weird crashes if I actually define a hook, though.

This is more flexible than a flag. With this we could compile
functions in a worker thread or something like that.

Having it separate from btb-compile-hook is important so that
(compile 'foo) works on bytecode functions even if auto compilation
is disabled, or defined in some other way.