bytecode FASLs: open file later
our with-open-file/close do not totally abort new files, so if
there's an error in the with-open-file body, you end up with an
empty file. This is kind of unfortunate for FASLs. We should fix
close, but in the meantime, we can just have the file compiler
delay opening the file. This should avoid an empty FASL being
produced in most cases, but it could still happen if there's some
error during %write-bytecode - which I think would have to be an
internal error/bug.