Fix bug allow save-lisp-and-die in successive sessions
This fixes #1370 - so I reopened that issue and this should close it.
The problem was that ObjectFile objects during snapshot_load would
pass their _MemoryBuffer to LLJIT and lose access to it.
With Lang Hames suggestion I make a shallow reference to the
_MemoryBuffer and pass that to LLJIT and keep the original.
Now we can load snapshots, do more stuff, and then save another
snapshot. Yay!