bytecode: redo bound declaration computation
Now bound declarations are stored in the BytecodeDebugVars along
with the frame index etc.

This requires a little more work by the bytecode compiler (to
figure out which declarations are bound) but nothing too onerous.
The advantage is that btb will be easier, and also it'll be easier
to get declaration information from a lexenv structure, e.g. for
Trucler.