btb refactor: fix issue with unreachable RETURNI
Code with infinite loops like (tagbody loop ... (go loop)) caused
a "bad rtype UNASSIGNED" error. This happened because we always
have a returni instruction in bytecode, and so we always generate
RETURNI BIR, even if the return isn't actually unreachable. The
returni block would be unknown to the flow order but still there
as the RETURNI slot of the function.

Deleting everything will wipe any unreachable blocks and properly
clean up their backpointers.