Delete bytecodep argument to lisp_bytecode_defun
It's a remnant of the bytecode wrapper switchover. And I want as
few calls to abort() as we can have. Ideally zero.

I think the BytecodeWrapper thing in general is obsolete, but
there'd be a lot more to delete for that, and I mostly just wanted
to ditch the abort for now.