Use templates for XEPs rather than copying XEPs around
This is basically in the service of making the function pointers
atomically mutable, which will make XEPs non-copyable.

Another thing to do would be to have the templates for most
types of global fun be static members, so they're not newly
constructed every time you make a global fun. They're not actually
heap allocated so it's probably not a big performance thing, but
nonetheless.