Remove LocalSimpleFun from simple fun hierarchy & rename
Function_O's SimpleFun slot is basically expected to match the
global simple fun convention, i.e. to have a function accepting
any number of boxed arguments and some functions for various
numbers of fixed arguments. That's not what these do or are
for, so having them be part of the hierarchy is just an invitation
to trouble.

I changed the name to CoreFun because "local function" might not
be true in a future where we have direct calls to these.