Introduce translate::make_from_object
This is for the case of doing a basic conversion of an object to
another object. Which is what I thought from_object was for
originally, but it also does some deeper magic with extents and
references and stuff... and magic should be kept to the wizards'
tower, so I'm trying to cut down on direct from_object usage.