Delete Cleavir evaluator
Never going to be fast. Also we don't actually need env:eval if
we go through bytecode for everything.

It might be a good idea to adopt some of what was in simple-eval
to the bytecode evaluator - picking out easy stuff like QUOTE or
PROGN - just because the bytecode compiler does still have some
avoidable overhead.