Fix Lisp FASL concatenation
We need to restart the count when we hit a new init-object-array,
as that makes sense and it's what loadltv expects. This can also
affect the length of indices. TODO: Test FASL concatenation, if
we're going to have the capability.

Note that this is the disltv FASL concatenation. During build we
usually use the C++ code to concatenate FASLs, specifically the
function core:link-fasl-files, which is not bugged.