Skip source compilation in make-foreign-caller
This is kind of awkward, and I'm not sure it's the right way to go
about it, but this means we don't need to go through the AST
compiler for this one thing. Since I'm planning on removing the AST
compiler that's pretty important.

Also removes the sham gcrootsinmodule since it is no longer
required by parseLinkGraph.

I skipped a commit from the old faso-cleanup branch that would
remove the Cleavir code walker. That's causing an error I don't
understand or want to deal with right now.