No more separate argument processing for local calls
Unifying this makes translate much less messy with all that
duplicate code, and makes it possible to local call anything,
including functions with &rest and/or &key parameters. Also means
we don't have to allocate a closure for local calls ever - that was
super dumb and I'm disappointed in myself for ever thinking it was
a good idea.

Argument parsing for local calls is moved into the caller in all
cases now. That may be a bit suboptimal for local functions that
are called in more than one place, since the parsing code is
duplicated. I really doubt that will be a big deal except _maybe_
for functions with &key, since key parsing is actually a fair bit
of code. But &key functions are already kind of doomed as far as
optimization goes, and local functions that are called in multiple
places are rare to begin with (only arising from flet/labels,
versus the much more common type checkers and m-v-bind lambdas) so
I'm not too worried. The Right Thing To Do long term is probably
to do as karlosz suggested years ago and represent XEPs explicitly
in the IR. A shorter term fix would be to create pseudo-XEPs for
these functions - "pseudo" in that they would parse arbitrary
arguments, but take the environment as individual arguments instead
of as an allocated closure.