Make backtraces a little better at optimized code
When LLVM optimization is on, it will often inline functions, and
in particular will inline core functions into their XEPs. The
DWARF information will reflect this inlining, so the address range
for a given return address may be of the inline function body
rather than that of the actual entry point. And that would break
backtraces.

This fix ignores the address ranges of the return address, instead
just checking the ranges of the object file's literal functions.
That works, but ideally we would use the DWARF information to give
us information about inlined functions. Also, with optimization on
a few tests on backtraces fail, so that also needs fixing.