analyzer: don't refer to *number-of-jobs*
This parallel search code seems to assume that it's being loaded
through the clasp-builder, but this is not necessarily the case,
and in fact isn't the case with how we usually use the analyzer
now - this code is rotten. Instead let's just use the previously
unused parameter.
The reference to *number-of-jobs* breaks with package locks on,
since without clasp-builder the symbol doesn't even exist, let
alone the variable.