remove classp check in setf-find-class
(setf find-class) already checks this, so having the underlying
function do it is redundant.
For my cross-compiling clasp experimentation I found it useful to
install a class that doesn't have a CPL yet, and clos:classp won't
work without a CPL. So that's another reason. The only reason to
check is sort of a sanity check to see that classes installed
before (setf find-class) is defined are okay, but if they're not
okay there's probably some deeper issue anyway.
Also this code dates back to when Clasp classes weren't instances.