Skip one comparison when comparing specializers
Since we already determined that spec1 /= spec2, and the CPL of
any class has that class as its first element, we can skip the
first element. Noticed this detail in the generalizers repo.
Also looked up the eql specializer rules and clarified that in a
comment. I don't know what "Beppe" means. That comment dates from
the initial commit of ECL over twenty years ago.