Use a concept for debug assert type cast
This is tricter - you need to be able to actually instantiate the
template anywhere there's an upcast, so e.g. the classes must be
fully defined. But I don't think that's a bad requirement honestly.

Doing it this way instead of static_assert means that the copy
constructor won't exist at all, rather than be made to exist and
then cause a compiler error. This also makes casting stricter.