Use atomics for signal+interrupt queueing
The old approach grabbed a lock in a signal handler, which I
suspect has caused me a few hangs over the years when I mashed
C-c C-c. I think the idea was that it's "safe" in that it used
atomics instead of AS-unsafe functions, but grabbing a lock is
not async-signal-safe no matter how you contrive it.

This approach to signals means that repeatedly signaling a thread
before it processes signals will result in the later signals being
dropped. But Linux does that for blocked signals anyway so there's
some precedent.

I also changed posix-interrupt to signal the interrupt immediately.
It's a little more messy looking, but it means we actually signal
the interrupt immediately like we ought to. Using INTERRUPT like I
was, I was hitting this odd issue where the first HUP (or whatever)
would get sent, and posix-interrupt would be called, but the
condition never got signaled or got delayed or something.
Signaling immediately fixed that.