Define booting processes to be active
and rename active->running internally, while keeping
process-active-p which is part of the interface and used in BT.

Booting is an implementation detail, but it's one we can't ignore.
A booting process is running but not ready for everything, most
notably interrupts. So the sensible thing to do (imo) is to hide
the detail by making interrupt-process hang a little bit.

The previous solution to this was to mark processes active when
they started booting, but that led to this interrupt problem.
I think booting might have been in there before and i screwed it
up. oh well, it works now and with proper synchronization.