Move interrupts disabled check to decision to handle
So, first off, fun fact, *interrupts-enabled* has been treated
backwards since it was introduced in 2017 - if it was T, interrupts
were actually DISabled. That's pretty cool.

Anyway, this commit. Up to now the interrupts system has been
pretty half-baked but the idea was essentially that disabling
interrupts disabled _asynchronous_ interrupts. You could still get
interrupts signaled by manually calling check-pending-interrupts,
or by having it be implicitly called whenever clasp happened to do
so (in practice, after any memory allocation). I don't think this
is very useful - interrupts can still happen whenever, except the
"whenever" is "whenever clasp feels like it" rather than "whenever
a signal arrives", and from the perspective of a user those are
not really all that different. If you have code that really truly
needs to not get interrupted, too bad for you, clasp might signal
an interrupt anyway.

So the new system, which I will write out explicitly somewhere as
I get through this round of changes, is that "disabling interrupts"
means that conditions of type INTERRUPT are not signalled. This
ideal isn't in place yet, and it'll be a little weird with
synchronous SEGV signals and stuff, but this commit moves us a bit
more towards that.