Disable interrupts in unwind-protect cleanups
this kludge took me a while to work through. anyway, doing this
seems subtle but important. If cleanups can be interrupted than
you can't meaningfully clean anything up. Got the idea from Racket.

Doing this with all our weird unwinding is kinda tricky. We also
have to make sure that interrupts can be reenabled within a
cleanup by the programmer with WITH-INTERRUPTS. hopefully the
programmer knows what they are doing in that event.