Move signal points from allocation to function calls
Signaling during allocation causes problems with the bytecode. The
VM does its own allocations and the VM state is not coherent
enough to be added to with frames from the signal handling code.
This was observed by varying the window size during build, which
triggered repeated SIGWINCH, but it happens with other signals as
well.

I've wanted to move these points from allocation to function calls
for a while. This SIGWINCH thing I accidentally caused makes the
issue more acute. Things seem basically okay like this - the system
is not obviously slower and remains responsive to keyboard
interrupts - but for really fast code we'll want to be able to
avoid these interrupt checks. And probably we should look at making
the check faster as well.