In snapshot load, don't bother with constructor
I don't see any reason not to just use fill directly. The commit
history doesn't make the justification clear. If it's supposed to
be for correctness, I don't see anything in cppreference that
would make going through a constructor more correct. I think the
correct thing would be for all objects to be trivially copyable,
but I don't think it's practically a problem even if they aren't.
(Which is good, because we have nontrivial destructors.)