Simplify object scanning code
The main thing is we have it only scan one object. The loop thing
is how MPS does it but does not seem to be common in GCs, insofar
as GCs have any commonality. If you want to do a loop you can
repeatedly call the scan function, since it returns a pointer just
past the object it scanned, now.

This also removes some skip calls in snapshot save and room. It
might speed them up a little, but probably not too much.