Delete some ctors+dtors in weak map code
replacing {} with default makes it trivial which is nice. As for
the dtors, from the boehm documentation, unregistration of a
disappearing link happens implicitly when it's deallocated, so I
don't think we need to bother with it. I also don't think these
destructors were ever actually called, so it's a good thing, too.