remove redundant test in hash table access
i'm reasonably sure this should never be tripped: no_key is not
usually available as a value, so nobody's storing it in hash tables
from Lisp. In C++ we only put it in the value as the default when
we're resizing a table, and in that case we make the key no_key as
well. So the only way you'd have a key that's not no_key and a
value that is no_key is I think if your hash table is thread unsafe
and you store a new key/value in it while another thread looks for
that key, and if that's what you're doing you're in crazytown to
begin with so the test is not going to help.