Fix scanning of weak key hash tables
We have been finding intermittent memory-test failures after
registering finalizers. The problem appears to be that the Buckets
for the values, which should be strong references, were instead
allocated boehm-atomically and therefore not scanned. This led to
live objects being collected, which causes lots of weird problems.
This only happened after registering finalizers because we use a
weak key hash table for finalizers internally.

The weak key hash table code is pretty brittle and I'm replacing it
in another branch.