Separate ISLLibraryHeader from rest of snapshot hierarchy
As mentioned in the comment, they're not stored in memory with the
rest of objects, so they don't need a marker word. Removing this
saves a tiny bit of space in snapshots, but more importantly
gets us closer to limiting the kinds to objects to clean up the
code. We could add the words back without restoring them to the
hierarchy if we need them for debugging at some point.