Preserve ephemerons in snapshot save/load
and make them traversable through ROOM

There may be an issue with the previous commit alone - it may not
allocate enough field infos/layouts in gc_boot.

The fact the memory walker uses T_O** fields is kind of dicey since
not all fields are actually pointers to pointers. We have the
hidden pointers in ephemerons, but also atomics in conses etc.