Use where tags instead of c++-stamp-max
If we give the user the ability to define packed structs, they may
be "C++ objects" as far as Lisp is concerned in that they use the
header stamp. So for discrimination to work the max is no good.

Just wanna get this out of the way now so I don't have inscrutable
problems later.