add function for computing object size
pretty dicey right now and I don't really want to advertise it to
users, but it's there. should help with debugging.

I'm seeing the size of a SimpleFun_O as 96 even though it only has
three fields? And the stamp dump also says the size is 88 instead.
I don't see this discrepancy with e.g. symbols. Odd.