Delete old object forwarding code
Some of it I don't understand the logic of. Why would we ever need
to know the size of a forwarding pointer? Even if we did, we could
just follow the pointer and see how big the forwarded object is.
And I don't see any purpose for the OBJECT_SIZE etc in obj_scan,
since we can just set forwarding pointers from the header.