Bump FASL version
There's a bit of awkwardness here - a while back I started doing
version 15 in Maclina, but paused work on Clasp's version 15 to do
other stuff. In the interim, yitzi put a version 15 into Clasp
that made different (and incompatible) changes. I think he asked
me first, but in any case it's messy. This wouldn't be acceptable
if the FASL format was like, used anywhere else, but it's still
experimental so not a huge deal.

I'm going to say that retroactively, 15 was Maclina's 15, and 16
is Maclina's 15 plus Clasp's 15.