Remove C package
In ECL, the C package is the compiler package. But it may be used
as a nickname by other systems and cause a collision. Also, nothing
actually seems to use cmpname, the only file using the package.
It was added in df8b9a5 for ASDF, but ASDF removed the use of
compute-init-name a few months earlier, and does not appear to
have ever used our compute-init-name, so I don't know what's up
with that.