Begin use of cross clasp
This adds an nclasp variant that builds a bytecode image of Clasp
from the host Lisp. This image contains most of Lisp, but not
compile-file, and is lacking some Clasp extensions.

Building nclasp-[gc] produces an nbase.fasl image file. This image
is not embedded in the executable or otherwise used unless you
explicitly specify it on the command line.