cross-clasp: Compiler macros
Ideally this would be defined earlier, but doing so is a real pain.
The whole point of this project is to _avoid_ bootstrapping
silliness, but unfortunately some will always be there, at least
while compilation and loading are based on side effects.

Also some of these compiler macros might need reevaluation.
Expanding MEMBER calls might not be very useful for bytecode. And
probably the VM should be able to handle APPLY natively (by making
it into MULTIPLE-VALUE-CALL) but that's a bigger change.