in FASL loader, use only global package lookup
I hit a situation where a package had a local nickname "CORE". So,
when the loader hit a CORE::*MAKE-SPECIAL - as in the Clasp CORE
package - it instead found the nicknamed package and things broke.

findPackageGlobal could be replicated by binding *package* to CL
or something, but that's more work to do less work.