cross-clasp: satiate properly
Now that I've set things up to always call generic functions,
rather than going through the "safe-" versions like in the old
code, we have to be sure to keep some core functions valid, so
that computing new discriminating functions is always possible.

This solution is more in line with beach's original ideas, as far
as I understand, and it's cleaner.