cross-clasp: compiled discriminating functions
Apparently we just don't have these hooked up to happen
automatically? Maybe we could throw that at the autocompilation
thread used for bytecode-to-bir.

Also I made it so you can compute an interpreted discriminator
even without a call history, if you want to for some perverse
reason. Just to avoid a weird exception.