fix updates for non-instances
This was causing problems since instance creators are doing
something weird- class-of (and core:instance-class) say they're
GENERALs, which is way too vague, but typep seems to work
correctly, i guess since their stamps are more specific?
Something to look into.

The instance-sig check in the original code should be unnecessary,
thanks to the changes to bootstrapping.