Fully commit to cross-clasp - no more old files
Hoping the histories work out correctly when this is merged back
into main. If not the history's there anyway.
If I ever have to do this again (god forbid) I should just warp
the original files, even if it breaks the build temporarily.

bytecode-machines is still in kernel2 since I need to make sure
generation of the vm-header works.