Pull compiled functions from literals instead of startup call
This has confused the hell out of me for years and I'm glad to fix
it. Ideally we would just install the constants from Lisp instead
of generating a function to do it in every single module, but baby
steps. This will hopefully make native code in bytecode FASLs
easier to accomplish.