Remove shady irbuilder in with-module
as far as I can tell this was not used for anything except a few
places it really shouldn't, relating to constants. I think what
happened is clasp-cleavir would make an int-to-ptr instruction
with the irbuilder, but it wouldn't actually go anywhere until
the constant was actually used later? Orrrrr something? Whatever,
now the constant is in fact a constant as it ought to be.

Didn't add a cmp: wrapper since i'm not convinced about those