clasp-cleavir: Keep XEPs out of literals vector if unneeded
They're only really needed in the literals vector if they're
actually used by the code - meaning, they're either called
or enclosed.

CoreFuns should never really need to be in the literals vector,
since they're not used for anything now and even if they were we
could just grab them from the SimpleCoreFun, but fixing this would
require running the analyzer and I'd like to hold off.