FASOs: move generators out-of-band
meaning that they do not go through the literals process. If a
generator appears as a literal in code something's wrong, and now
it will be an error (no MAKE-LOAD-FORM) instead of stumbling along.

This also further distinguishes FASO building from runtime native
code compilation, which is my actual goal here.