auto-compile whole modules at once
This makes things a bit more transparent - the identities of
inner functions don't change now. And it should make compiling non
global definitions fine. There might be a rare race condition if an
inner function with a nonlocal return is compiled and called
before its destination is. Not sure, seems unlikely. In any case
I'd eventually like to make it possible for native functions to
return to bytecode, and vice versa.