Allow serializing some custom hash tables
This is a kludge and I don't like it, so I'm not sure I want to
document this at least for now, but oh well. It works if the
functions are global and have their names properly attached, and
if the functions are defined the same in the loading image.
Things will break weirdly if they're not.