Fix fencepost error in b2b
The LONG byte of instructions was not counted for opip, i.e. opip
pointed at the byte _after_ the long prefix. Since jumps point _at_
the long prefix, this resulted in any blocks that happened to start
with a long instruction being treated as unreachable, resulting in
miscompilation.

Test case was generated code consisting of 255 (or so) INTERN calls
with different names, followed by a loop. A block in the loop would
begin with a long fdefinition, and from there the problems begin.